preserve session start time when we handleNewVoter
The first UserInfo is created when we get the first message, but
handleNewVoter *also* creates a new one.  (This is I think because there
are multiple callers... and one of them doesn't have a UserInfo.)

Pass in our original userInfo (if we have one) so that we don't ignore
the part of the message history before they texted HELPLINE.

Also, fix the start time: (1) correct for TZ (ugh why don't we use UTC
everywhere?) and (2) pad by 10 seconds so that we capture the most recent
incoming message.