stale thread: only link to new thread if there is one
If you post to an old thread, it will tell you it is stale and link to
the new one.  Do not try to link if there is no new thread.

This happens if a user texts something other than helpline, triggering
the old thread to close, but doesn't say HELPLINE so a new thread isn't
opened.