route a voter => volunteerEngaged
Otherwise, a fresh voter could get routed by a volunteer to a state
channel, and if the next message is from the voter it will get
interpreted as a state attempt.

Don't put this in routeVoterToSlackChannel() since one caller is
automated and setting this value doesn't make sense in that case.