getLatestVoterStatus: ignore SPAM/REFUSED/REJOIN
This is used to populate the voter status dropdown, and should ignore
any intervening REFUSED -> REJOIN cycles (or even REFUSED), since those
aren't valid drop-down options.  These special statuses are only used to
tell whether a voter is blocked or not but should not affect what stage
of the voting process the voter is in.