Add !state <statename> command
This is simpler than using !route because you just specify the state
name (full name or abbreviation) and the helpline routes you (randomly)
to an appropriate pod.  No need to spell out the full name, and no need
to pick which pod for the state, and no need to remember which states have
a dedicated pod vs a regional or multi-state one.

Also, this updates UserInfo.stateName, which means that the introduction
of the voter into the channel correctly indicates which state the voter
is from.  (This is not updated when using !route directly.)