Refresh slack blocks from database instead of modifying in place
The old strategy of fetching blocks from the slack API, modifying them,
and then pushing them back to slack is fragile:

- It's hard to follow the update logic
- It is even harder to write code that deals with different iterations
of the structure of the slack blocks, which means it is very difficult to
add or remove items to the panel.

Instead, rebuild the panel entirely from scratch whenever we need to
refresh. This can be done with quick queries to