Avoid fetching blocks when closing a voter panel
In most cases we already had the userInfo; in the reset demo modal case,
we fetch it.

This has the side-effect that an old voter panel that hasn't refreshed
will get new initial text as it is closed.  This is no different than
if the voter were updated in some other way (e.g., voter status changed)
and then reset.